feat: remove `copy_db_to_local_drive()`
This function is dangerous, because if two jobs end up running on the
same node, the one that started first could clean up the database before
the other is done using it.  Fixing this would require adding some sort
of reference counting mechanism.  For now, better to just remove it.