feat: restart data loader processes each epoch
My atompaint datasets used to have state, so it was necessary to prevent
the data loader processes from restarting during the training run.
However, this approach turned out to be incompatible with requeueing, so
I had to rewrite my datasets to be stateless.

Now, I think that using the same data loader processes for the whole
training run can be harmful.  I had a training run fail due to an
out-of-memory error after 18 epochs; memory use shouldn't increase after
the first epoch.  I suspect this is due to some sort of caching, and
that a robust way to prevent similar problems in the future is to
regularly restart the data loaders.